Update Schedule model to allow null fieldofStudy and updated test
