+++ lib/logic/settings/settings_cubit.freezed.dart
@@ -169,7 +169,7 @@ class _$SettingsStateImpl extends _SettingsState {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$SettingsStateImpl &&
